// src/shared/types/PodiumScoreboard.ts
export type TPodiumScoreboard = {
  id: number
  universityName: string
  basketball_points: number
  football_points: number,
  pingpong_points:number,
  badminton_points: number,
  athletics_points: number,
  tradition_game_points: number;
  esport_point: number;
}

        // {
        //     "id": 4,
        //     "universityName": "KMUTNB_PR",
        //     "basketball_points": 14,
        //     "football_points": 11,
        //     "pingpong_points": 7,
        //     "badminton_points": 10,
        //     "athletics_points": 12
        // }
        